+++ src/app/components/space-automation/atm-workflow-executions-table.js
@@ -73,18 +73,21 @@ export default Component.extend(I18n, {
       case 'waiting':
         return [
           'name',
+          'inventory',
           'scheduledAt',
           'status',
         ];
       case 'ongoing':
         return [
           'name',
+          'inventory',
           'startedAt',
           'status',
         ];
       case 'ended':
         return [
           'name',
+          'inventory',
           'startedAt',
           'finishedAt',
           'status',
@@ -122,7 +125,7 @@ export default Component.extend(I18n, {
         return atmWorkflowExecutionSummaryIndex(record, phase);
       },
       startIndex: 0,
-      endIndex: 20,
+      endIndex: 50,
       indexMargin: 10,
     });
   }),