+++ src/app/services/mock-backend.js
@@ -1087,6 +1087,8 @@ export default Service.extend({
    * @returns {Promise<Array<Model>>}
    */
   async createAtmWorkflowExecutionRecords(store) {
+    const space = this.get('entityRecords.space.0');
+    const atmInventories = this.get('entityRecords.atmInventory');
     const atmWorkflowSchemas = this.get('entityRecords.atmWorkflowSchema');
     const atmWorkflowSchemasCount = get(atmWorkflowSchemas, 'length');
     const timestamp = Math.floor(Date.now() / 1000);
@@ -1100,6 +1102,7 @@ export default Service.extend({
       const phaseIndex = atmWorkflowExecutionPhases.indexOf(phase);
       const executionsGroup = await allFulfilled(
         _.range(numberOfAtmWorkflowExecutions).map(async (i) => {
+          const atmInventory = atmInventories[i % atmInventories.length];
           const scheduleTime = phaseIndex >= waitingPhaseIndex ?
             timestamp + i * 3600 : null;
           const startTime = phaseIndex >= ongoingPhaseIndex ?
@@ -1172,6 +1175,7 @@ export default Service.extend({
             startTime,
             finishTime,
             atmWorkflowSchemaSnapshot,
+            space,
           }).save();
           atmWorkflowExecutions.push(atmWorkflowExecution);
           return await store.createRecord('atmWorkflowExecutionSummary', {
@@ -1187,6 +1191,7 @@ export default Service.extend({
             startTime,
             finishTime,
             atmWorkflowExecution,
+            atmInventory,
           }).save();
         })
       );